ppo and reward model update
