+++ optimum/habana/trl/__init__.py
@@ -2,4 +2,5 @@ from .models.modeling_base import adapt_PreTrainedModelWrapper_to_gaudi
 from .trainer.dpo_trainer import GaudiDPOTrainer
 from .trainer.ppo_config import GaudiPPOConfig
 from .trainer.ppo_trainer import GaudiPPOTrainer
+from .trainer.reward_trainer import GaudiRewardTrainer, RewardDataCollatorWithPadding
 from .trainer.sft_trainer import GaudiSFTTrainer
