+++ optimum/habana/trl/models/__init__.py
@@ -1,5 +1,3 @@
-# flake8: noqa
-
 # Copyright 2022 The HuggingFace Team. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
