+++ optimum/habana/trl/trainer/__init__.py
@@ -21,3 +21,4 @@ from .sft_trainer import GaudiSFTTrainer
 from .dpo_trainer import GaudiDPOTrainer
 from .ppo_config import GaudiPPOConfig
 from .ppo_trainer import GaudiPPOTrainer
+from .reward_trainer import GaudiRewardTrainer, RewardDataCollatorWithPadding
